pyproject.toml
[project]
name = "brain-ops"
version = "0.2.0"
description = "Operational backend for an Obsidian-based second brain"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "typer>=0.12.0",
  "pydantic>=2.7.0",
  "pyyaml>=6.0.1",
  "rich>=13.7.0",
]

[project.scripts]
brain = "brain_ops.cli:app"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"
